feat: improve accuracy at lower wattages & increase telemetry interval from 60s->10s

+++ src/power/PowerModule.hpp
@@ -13,7 +13,7 @@ private:
     EnergyMonitor energyMonitor;
 
 public:
-    int getPowerUsageWatts(); // Returns the measured power in Watts
+    double getPowerUsageWatts(); // Returns the measured power in Watts
 
     static PowerModule &getInstance();
